impl renderer and training code
The renderer will now compute the 2D covariances, evaluate the 2D
gaussians, and obtain alpha values. This is used to obtain
transmittances, and this is all used to obtain the final renderered
outputs (depth maps, RGB images, masks).

The training code is also implemented -- learning rates are set
for the various params, the params are registered with the optimizer,
and we requires_grad for all of them.